feat: split screen
